refactor(form): move the date and time pickers into primitives
